Docs: cleanup.
